+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- `authService.getCurrentFormsAppUser()` to call API regardless of `isSAMLUser`
+
 ### Fixed
 
 - autosave data not being removed after saving a draft offline